Review fixes: move provenance tests to acceptance, clean imports
- Move PROVENANCE_ACCEPTANCE_TESTS from field_validation to acceptance file
- Remove unused assertNotError import from validation file
- Clarify bulkWrite exclusion rationale in utils comment
