+++ documentdb_tests/compatibility/tests/core/query_and_write/write_concern/test_write_concern_acceptance.py
@@ -199,6 +199,38 @@ def test_write_concern_wtimeout_edge_cases(collection, test: CommandTestCase):
     assertNotError(result, msg=test.msg)
 
 
+_PROVENANCE_VALUES = [
+    ("clientSupplied", "clientSupplied"),
+    ("implicitDefault", "implicitDefault"),
+    ("customDefault", "customDefault"),
+    ("getLastErrorDefaults", "getLastErrorDefaults"),
+    ("null", None),
+]
+
+# Property [Provenance Acceptance]: writeConcern accepts provenance sub-field.
+PROVENANCE_ACCEPTANCE_TESTS: list[CommandTestCase] = [
+    CommandTestCase(
+        f"{cmd}_accepts_provenance_{val_name}",
+        docs=[{"_id": 1}],
+        command=lambda ctx, _prov=value, _cmd=cmd: build_cmd(
+            _cmd, ctx, {"w": 1, "provenance": _prov}
+        ),
+        msg=f"{cmd} should accept provenance:'{value}'.",
+    )
+    for cmd in WRITE_COMMANDS
+    for val_name, value in _PROVENANCE_VALUES
+]
+
+
+@pytest.mark.parametrize("test", pytest_params(PROVENANCE_ACCEPTANCE_TESTS))
+def test_write_concern_provenance_accepted(collection, test: CommandTestCase):
+    """Test writeConcern accepts provenance sub-field values."""
+    collection = test.prepare(collection.database, collection)
+    ctx = CommandContext.from_collection(collection)
+    result = execute_command(collection, test.build_command(ctx))
+    assertNotError(result, msg=test.msg)
+
+
 def test_write_concern_null_equivalent_to_omitted(collection):
     """Test writeConcern null produces same success as omitting writeConcern."""
     collection.insert_one({"_id": 1, "a": 0})
